no longer using imagejpeg to output avatar to browser

+++ actions/crop.php
@@ -6,17 +6,20 @@ $action = postVarClean('action');
 $u_id = postVarClean('u_id') != "" ? postVarClean('u_id') : getVarClean('u_id');
 
 
-$big_filename = $DOCUMENTS_PATH . $u_id . '_big.jpg';
-$small_filename = $DOCUMENTS_PATH . $u_id . '_small.jpg';
+$big_filename = dirname(__FILE__) . '/../' . $AVATARS_PATH . $u_id . '_big.jpg';
+$small_filename = dirname(__FILE) . '/../' . $AVATARS_PATH . $u_id . '_small.jpg';
 
 //TODO: 
 // file size
-// file type
+// file type (convert to jpg?)
 // use success/failure msgs 
 
 if ($action == 'upload') {
 	if(move_uploaded_file($_FILES['uploadedfile']['tmp_name'], $big_filename)) {
-		unlink($small_filename);
+		copy($big_filename, dirname(__FILE__) . '/../' . $AVATARS_PATH . $u_id . '_big2.jpg');
+		if (file_exists($small_filename)) {
+			unlink($small_filename);
+		}
 		echo 'Pic upload success';
 	} else{
 	  echo "There was an error uploading the file, please try again!";
@@ -36,21 +39,41 @@ if ($action == 'upload') {
 
 	imagecopyresampled($dst_r, $img_r, 0, 0, $x, $y, $targ_w, $targ_h, $w, $h);
 
-	$big_filename = $DOCUMENTS_PATH . $u_id . '.jpg';
 	if(imagejpeg($dst_r, $small_filename, $jpeg_quality)) {
 		echo 'Success';
 	} else {
 		echo 'There was an error saving.';
 	}
+
+	imagedestroy($img_r);
+	imagedestroy($dst_r);
 } else {
+
+/*
 	$type = getVarClean('type');
-	$filename = $big_filename;
-	if (file_exists($filename)) {
-		$img = imagecreatefromjpeg($filename);		
-		header('Content-type: image/jpeg');
-		imagejpeg($img);
-		imagedestroy($img);
+	$a = getVarClean('a');
+	if ($type == 'cropped') {
+		$filename = $small_filename;
+	} else if ($a == 'a'){
+		$filename = $bigg;
+	} else {
+		$filename = $big_filename;
 	}
+	if(file_exists($filename)) {
+		$img = imagecreatefromjpeg($filename);		
+	} else {
+		$img = imagecreatefromjpeg('../images/default.jpg');
+	}	
+	
+	
+	header('Content-type: image/jpeg');
+//	header('Expires: Thu, 01 Dec 1994 16:00:00 GMT');
+//	header('Cache-Control: no-cache, no-store');
+//	header('Last-Modified: '.gmdate('D, d M Y H:i:s', time()).' GMT'); 
+
+	imagejpeg($img);
+	imagedestroy($img);
+*/
 }
